% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_content.R
\name{article_content}
\alias{article_content}
\title{Article Content Scraper (Papua New Guinea News Websites)}
\arguments{
\item{url}{Takes a character url as input.}
}
\value{
Returns an object of class 'character'
}
\description{
Scrapes news article contents from either post-courier, loop png or the national news websites.
}
\examples{
url <- "https://www.looppng.com/business/fiscal-stability-agreement-p\%E2\%80\%99nyang-signed-125267" 
content <- article_content(url);
}
