% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_business_news_scraper.R
\name{business_lp}
\alias{business_lp}
\title{Business News Scraper (Loop PNG)}
\arguments{
\item{pages}{Takes an integer value as input. This allows the script to page through the website.}
}
\value{
Returns an object of class 'tibble'
}
\description{
Scrapes business news articles from loop png news website
}
\examples{
df <- business_lp(pages = 1);
df2 <- business_lp(1);
}
