% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_business_news_scraper.R
\name{business_na}
\alias{business_na}
\title{Business News Scraper (The National News Paper)}
\arguments{
\item{pages}{Takes an integer value as input. This allows the script to page through the website.}
}
\value{
Returns an object of class 'tibble'
}
\description{
Scrapes business news articles from the national news website
}
\examples{
df <- business_na(pages = 1);
df2 <- business_na(1);
}
