% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_business_news_scraper.R
\name{business_pc}
\alias{business_pc}
\title{Business News Scraper}
\arguments{
\item{pages}{Takes an integer value as input. This allows the script to page through the website.}
}
\value{
Returns an object of class 'tibble'
}
\description{
Scrapes business news articles from the post-courier news website
}
\examples{
df <- business_pc(pages = 1);
df2 <- business_pc(1);
}
