% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_feature_news_scraper.R
\name{feature_pc}
\alias{feature_pc}
\title{Feature News Scraper}
\arguments{
\item{pages}{Takes an integer value as input. This allows the script to page through the website.}
}
\value{
Returns an object of class 'tibble'
}
\description{
Scrapes featured news articles from the post-courier news website.
}
\examples{
df <- feature_pc(pages = 1);
df2 <- feature_pc(1);
}
