% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_national_news_scraper.R
\name{national_lp}
\alias{national_lp}
\title{National News Scraper (Loop PNG)}
\arguments{
\item{pages}{Takes an integer value as input. This allows the script to page through the website.}
}
\value{
Returns an object of class 'tibble'
}
\description{
Scrapes national news articles from loop png website
}
\examples{
df <- national_lp(pages = 1);
df2 <- national_lp(1);
}
