% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_national_news_scraper.R
\name{national_na}
\alias{national_na}
\title{National News Scraper (The National News Paper)}
\arguments{
\item{pages}{Takes an integer value as input. This allows the script to page through the website.}
}
\value{
Returns an object of class 'tibble'
}
\description{
Scrapes national news articles from the National News website
}
\examples{
df <- national_na(pages = 1);
df2 <- national_na(1);
}
