% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_national_news_scraper.R
\name{national_pc}
\alias{national_pc}
\title{National News Scraper}
\arguments{
\item{pages}{Takes an integer value as input. This allows the script to page through the website.}
}
\value{
Returns an object of class 'tibble'
}
\description{
Scrapes national news articles from the post-courier news website
}
\examples{
df <- national_pc(pages = 1);
df2 <- national_pc(1);
}
