% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_validate.R
\name{page_validate}
\alias{page_validate}
\title{Page Validate}
\arguments{
\item{pages}{Takes an integer value as input needing validation.}

\item{url}{A web address as a character string or an object of character string.}
}
\value{
If validation passes, function returns an integer or returns NULL if fails.
}
\description{
Function that validates the page argument in the pngnewsR package functions.
}
\examples{
url <- "https://www.postcourier.com.pg/new-category/features-news/";
validation_1 <- page_validate(1000000, url);
validation_2 <- page_validate(1, url);
}
