% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_sport_news_scraper.R
\name{sport_lp}
\alias{sport_lp}
\title{Sport News Scraper (Loop PNG)}
\arguments{
\item{pages}{Takes an integer value as input. This allows the script to page through the website.}
}
\value{
Returns an object of class 'tibble'
}
\description{
Scrapes sports news articles from loop png website
}
\examples{
df <- sport_lp(pages = 1);
df2 <- sport_lp(1);
}
