% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_sport_news_scraper.R
\name{sport_pc}
\alias{sport_pc}
\title{Sport News Scraper}
\arguments{
\item{pages}{Takes an integer value as input. This allows the script to page through the website.}
}
\value{
Returns an object of class 'tibble'
}
\description{
Scrapes sport news articles from the post-courier news website
}
\examples{
df <- sport_pc(pages = 1);
df2 <- sport_pc(1);
}
