% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topstories_news_scraper.R
\name{topstories}
\alias{topstories}
\title{Top Stories News Scraper}
\arguments{
\item{pages}{Takes an integer value as input. This allows the script to page through the website.}
}
\value{
Returns an object of class 'tibble'
}
\description{
Scrapes top stories news articles from the post-courier news website
}
\examples{
df <- topstories(pages = 1);
df2 <- topstories(1);
}
