% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_news_scraper.R
\name{world}
\alias{world}
\title{World News Scraper}
\arguments{
\item{pages}{Takes an integer value as input. This allows the script to page through the website.}
}
\value{
Returns an object of class 'tibble'
}
\description{
Scrapes world news articles from the post-courier news website
}
\examples{
df <- world(pages = 1);
df2 <- world(1);
}
